Rename graphql.dictionary-service to graphql-dictionary-service. Rename rest.dictionary-service to rest-dictionary-service.
