+++ packages/sitecore-jss/src/i18n/rest-dictionary-service.ts
