+++ packages/sitecore-jss/src/index.ts
@@ -46,5 +46,5 @@ export { DictionaryPhrases, DictionaryService } from './i18n/dictionary-service'
 export {
   GraphQLDictionaryService,
   GraphQLDictionaryServiceConfig,
-} from './i18n/graphql.dictionary-service';
-export { RestDictionaryService, RestDictionaryServiceConfig } from './i18n/rest.dictionary-service';
+} from './i18n/graphql-dictionary-service';
+export { RestDictionaryService, RestDictionaryServiceConfig } from './i18n/rest-dictionary-service';
